+++ extensions/punto/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Punto Changelog
 
+## [Support Ilya Birman Typography] - 2024-01-05
+
+- Added support Ilya Birman Typography Layout
+- Update keyboardSwitcher v2.0.0
+
 ## [Initial Version] - 2022-12-29
 
 - Added layout switching EN-RU in both directions